fix: migrate reorder-links to zustand
