+++ src/lib/links/index.ts
@@ -1,2 +1,3 @@
 export * from "./add-link";
+export * from "./reorder-links";
 export * from "./update-link";
